add vercel build script
