+++ .gitignore
@@ -177,3 +177,5 @@ dist
 .aider*
 
 tmp
+.vercel
+public
