+++ dev-server.ts
@@ -1,5 +1,4 @@
-import { getSvg } from "./drawing/getSvg"
-import symbols from "./symbols"
+import { generateWebPage } from "./scripts/lib/generate-web-page"
 
 console.log(`Serving on http://localhost:3077`)
 Bun.serve({
@@ -8,60 +7,7 @@ Bun.serve({
     const url = new URL(req.url)
 
     if (url.pathname === "/") {
-      const symbolEntries = Object.entries(symbols).sort((a, b) =>
-        a[0].localeCompare(b[0]),
-      )
-
-      const svgGrid = symbolEntries
-        .map(([name, symbol]) => {
-          const svg = getSvg(symbol, { width: 150, height: 150, debug: true })
-          return `
-          <div class="symbol-container" style="padding-bottom:28px;">
-            <div style="font-size: 12px;word-break: break-all; text-align: left;padding-bottom: 16px;">${name}</div>
-            ${svg}
-          </div>
-        `
-        })
-        .join("")
-
-      const html = `
-        <!DOCTYPE html>
-        <html lang="en">
-        <head>
-          <meta charset="UTF-8">
-          <meta name="viewport" content="width=device-width, initial-scale=1.0">
-          <title>Schematic Symbols Grid</title>
-          <style>
-            body {
-              font-family: Arial, sans-serif;
-              margin: 0;
-              padding: 20px;
-            }
-            .grid {
-              display: grid;
-              grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-              gap: 20px;
-            }
-            .symbol-container {
-              border: 1px solid #ccc;
-              padding: 10px;
-              text-align: center;
-            }
-            svg {
-              max-width: 100%;
-              height: auto;
-            }
-          </style>
-        </head>
-        <body>
-          <h1>Schematic Symbols</h1>
-          <div class="grid">
-            ${svgGrid}
-          </div>
-        </body>
-        </html>
-      `
-
+      const html = generateWebPage()
       return new Response(html, {
         headers: { "Content-Type": "text/html" },
       })
