+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "start": "bun run build && bun --hot run dev-server.ts",
     "build": "rm symbols/index.ts && bunx make-vfs ./symbols --content-format import-default --outfile ./symbols/index.ts",
-    "generate": "bun run scripts/generate-symbols-from-asset-svgs.ts"
+    "generate": "bun run scripts/generate-symbols-from-asset-svgs.ts",
+    "vercel-build": "bun run ./scripts/build.ts"
   },
   "devDependencies": {
     "@types/bun": "latest",
